cleaned duplication created by adding react-native-firebase checked it is not broken at yarn start
