--- RNFirebase/app.json
@@ -1,4 +1,0 @@
-{
-  "name": "RNFirebase",
-  "displayName": "RNFirebase"
-}
