+++ android/app/src/main/java/com/rnfirebase/MainApplication.java
