+++ android/settings.gradle
