+++ ios/RNFirebase-tvOS/Info.plist
