+++ ios/RNFirebase-tvOSTests/Info.plist
