+++ ios/RNFirebase/AppDelegate.h
