+++ ios/RNFirebase/AppDelegate.m
