+++ ios/RNFirebase/Images.xcassets/AppIcon.appiconset/Contents.json
