+++ ios/RNFirebase/Images.xcassets/Contents.json
