+++ ios/RNFirebase/Info.plist
