+++ ios/RNFirebaseTests/Info.plist
