+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "storymap",
+  "version": "0.0.1",
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
@@ -15,15 +16,19 @@
     "@expo/samples": "2.1.1",
     "expo": "^31.0.2",
     "node": "^11.2.0",
-    "react": "16.5.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-31.0.0.tar.gz",
+    "react": "16.6.1",
+    "react-native": "0.57.7",
     "react-native-cli": "^2.0.1",
     "react-native-firebase": "^5.1.1",
     "react-navigation": "^2.18.2"
   },
   "devDependencies": {
     "babel-preset-expo": "^5.0.0",
-    "jest-expo": "^31.0.0"
+    "jest-expo": "^31.0.0",
+    "babel-jest": "23.6.0",
+    "jest": "23.6.0",
+    "metro-react-native-babel-preset": "0.50.0",
+    "react-test-renderer": "16.6.1"
   },
   "private": true
 }
